move root and diff to separate files
